requirements.txt
Pillow>=10.0.0
luma.oled>=3.14.0
psutil>=5.9.0
evdev>=1.7.0
RPi.GPIO>=0.7.1
